Complete usable version of the rewrite
This commit completes a first usable version of the rewrite which will
later function as a module for the bot script.
This version uses the UNIX select() call for efficient usage and adds a
queue for sending messages. This enables to prepare messages for
sending prior to entering the main loop.
A rudimentary documentation is included in form of Python Docstring
inlined into the source code.